Detect line endings in dependency files.
When running Windows executables on a Unix system under Wine compilers
can emit dependency files with Windows-style paths. This causes ninja
to force rebuilds as it cannot find the resulting files.

This commit adds logic to the dependency file parser to detect the line
endings of the file. This can be used as a test to determine which deps
need their paths converted.