Hacky support for rewriting Wine paths.
This works, but is ugly. It modifies unowned `StringPiece`s in-place.

When running Windows executables on a Unix system under Wine compilers
can emit dependency files with Windows-style paths. This causes ninja
to force rebuilds as it cannot find the resulting files.

This commit uses the line endings of a file to determine if directory
separators should be converted. If the line on Unix are DOS-style or
on Windows are Unix-style then all inputs in a compiler-generated
dependency file are rewritten to the native directory separator. Paths
in a file with mixed line endings are left untouched.